add fastapi.staticfile and put files in static directory
